Refactored delete dialogs to different file
